requirements.txt
google-cloud-bigquery>=3.11.4
google-auth>=2.22.0
python-dotenv>=1.0.0
requests>=2.31.0
alembic>=1.13.1
sqlalchemy>=2.0.27
pandas>=2.0.0
plotly>=5.18.0
